[CrankshaftMove] periodic image added.
